Move RPC API to qubic_rpc::server
